:recycle: divide and concur
